修改打印方式，在打印信息前面添加运行时间 Signed-off-by: liudonghua7914 <657226657@qq.com>

+++ Common/Inc/def.h
@@ -9,9 +9,9 @@ typedef unsigned char  BYTE;
 typedef unsigned int   UINT;
 typedef unsigned short WORD;
 typedef unsigned short UINT16;
-typedef unsigned long  UINT32;
-typedef unsigned long  DWORD;
-typedef unsigned long  ULONG;
+typedef unsigned int  UINT32;
+typedef unsigned int  DWORD;
+typedef unsigned int  ULONG;
 
 typedef unsigned char  U8;
 typedef unsigned char  u8;
@@ -22,11 +22,11 @@ typedef unsigned int  u32;
 
 typedef unsigned char  Uint8;
 typedef unsigned short Uint16;
-typedef unsigned long  Uint32;
+typedef unsigned int  Uint32;
 
 typedef unsigned char  uint8;
 typedef unsigned short uint16;
-typedef unsigned long  uint32;
+typedef unsigned int  uint32;
 
 typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
